mudando tipo da quantidade para ficar igual ao tipo da quantidade do método adicionarAlimento de refeicaoService para evitar problemas

+++ src/main/java/br/inatel/nutriPlan/controllers/RefeicaoController.java
@@ -86,7 +86,7 @@ public class RefeicaoController {
     @PostMapping("/{refeicaoId}/alimentos")
     public ResponseEntity<Refeicao> adicionarAlimento(@PathVariable long refeicaoId,
                                                       @RequestParam long alimentoId,
-                                                      @RequestParam int quantidade) {
+                                                      @RequestParam double quantidade) {
         Refeicao refeicao = refeicaoService.adicionarAlimento(refeicaoId, alimentoId,quantidade);
         return ResponseEntity.status(HttpStatus.CREATED).body(refeicao);
     }
